# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de controle financeiro pessoal com autenticação JWT",
        "title": "Imperium API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tron/decisions/{did}/resolve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "CIO resolves a pending decision by choosing an option",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-decisions"
                ],
                "summary": "Resolve a decision",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Decision ID",
                        "name": "did",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Resolution",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResolveDecisionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TronDecision"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/tron/projects/{id}/agents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current status and metrics of all TRON agents",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-agents"
                ],
                "summary": "Get status of all agents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TronAgentsStatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tron/projects/{id}/agents/run": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually triggers a full agent cycle for the project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-agents"
                ],
                "summary": "Manually trigger an agent cycle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Cycle started",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tron/projects/{id}/decisions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all decisions requiring CIO input",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-decisions"
                ],
                "summary": "List decisions for a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending, approved, rejected)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TronDecisionListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ResolveDecisionRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "option_id": {
                    "type": "string"
                }
            }
        },
...
        "models.TronAgentStatus": {
            "type": "object",
            "properties": {
                "agent_type": {
                    "$ref": "#/definitions/models.TronAgentType"
                },
                "avg_duration_ms": {
                    "type": "integer"
                },
                "failed_runs": {
                    "type": "integer"
                },
                "is_running": {
                    "type": "boolean"
                },
                "last_run_at": {
                    "type": "string"
                },
                "last_run_success": {
                    "type": "boolean"
                },
                "successful_runs": {
                    "type": "integer"
                },
                "total_cost_usd": {
                    "type": "number"
                },
                "total_runs": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
        "models.TronAgentType": {
            "type": "string",
            "enum": [
                "orchestrator",
                "board",
                "pm",
                "dev",
                "qa",
                "integration"
            ],
            "x-enum-varnames": [
                "AgentTypeOrchestrator",
                "AgentTypeBoard",
                "AgentTypePM",
                "AgentTypeDev",
                "AgentTypeQA",
                "AgentTypeIntegration"
            ]
        },
        "models.TronAgentsStatusResponse": {
            "type": "object",
            "properties": {
                "agents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TronAgentStatus"
                    }
                },
                "cycle_running": {
                    "type": "boolean"
                },
                "next_cycle_at": {
                    "type": "string"
                }
            }
        },
...
        "models.TronDecision": {
            "type": "object",
            "properties": {
                "agent_type": {
                    "$ref": "#/definitions/models.TronAgentType"
                },
                "chosen_option": {
                    "type": "string"
                },
                "context": {
                    "description": "Additional context",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "default_option": {
                    "description": "Which option to use on timeout",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "$ref": "#/definitions/models.TronDecisionLevel"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TronDecisionOption"
                    }
                },
                "project_id": {
                    "type": "string"
                },
                "repo_id": {
                    "type": "string"
                },
                "resolved_at": {
                    "type": "string"
                },
                "resolved_by": {
                    "description": "\"user\" or \"timeout\" or \"auto\"",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.TronDecisionStatus"
                },
                "task_id": {
                    "type": "string"
                },
                "timeout_at": {
                    "description": "When to auto-resolve",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.TronDecisionLevel": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-comments": {
                "DecisionLevelCritical": "Critical, needs attention",
                "DecisionLevelInfo": "Just informational",
                "DecisionLevelNormal": "Normal decision"
            },
            "x-enum-descriptions": [
                "Just informational",
                "Normal decision",
                "Critical, needs attention"
            ],
            "x-enum-varnames": [
                "DecisionLevelInfo",
                "DecisionLevelNormal",
                "DecisionLevelCritical"
            ]
        },
        "models.TronDecisionListResponse": {
            "type": "object",
            "properties": {
                "decisions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TronDecision"
                    }
                },
                "pending": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.TronDecisionOption": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "impact": {
                    "description": "What happens if chosen",
                    "type": "string"
                },
                "is_default": {
                    "description": "Used if timeout",
                    "type": "boolean"
                },
                "label": {
                    "type": "string"
                }
            }
        },
        "models.TronDecisionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected",
                "timeout",
                "auto"
            ],
            "x-enum-comments": {
                "DecisionStatusAuto": "Automatically resolved",
                "DecisionStatusTimeout": "Auto-resolved by default"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "Auto-resolved by default",
                "Automatically resolved"
            ],
            "x-enum-varnames": [
                "DecisionStatusPending",
                "DecisionStatusApproved",
                "DecisionStatusRejected",
                "DecisionStatusTimeout",
                "DecisionStatusAuto"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite: Bearer {seu_token_aqui}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}